requirements.txt
auto_mix_prep==0.2.0
cdsapi==0.7.6
matplotlib==3.10.6
numpy==2.3.3
Office365-REST-Python-Client==2.6.1
pandas==2.3.2
PyYAML==6.0.3
scikit_learn==1.7.2
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
xarray==2025.9.0
plotly==6.3.0
scipy==1.16.2
seaborn==0.13.2
einops==0.8.1
huggingface_hub==0.35.1
Requests==2.32.5
timm==1.0.207
cfgrib==0.8.4.5
